backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for my roleplaying DB"
authors = ["Jacob Ràfols <ovnijacob@gmail.com>"]
license = "Ask Me later, I don't know yet"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
graphene-django = "^3.0.0"
django-filter = "^22.1"
hashids = "^1.3.1"
django-cors-headers = "^3.13.0"
django-graphql-jwt = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
